+++ interfaces/HO-Portal/src/app/program/program-people-affected/program-people-affected.component.ts
@@ -10,7 +10,7 @@ import { BulkActionsService } from 'src/app/services/bulk-actions.service';
 import { ProgramsServiceApiService } from 'src/app/services/programs-service-api.service';
 import { formatPhoneNumber } from 'src/app/shared/format-phone-number';
 import { environment } from 'src/environments/environment';
-import { PaymentErrorPopupComponent } from '../payment-error-popup/payment-error-popup.component';
+import { PaymentStatusPopupComponent } from '../payment-status-popup/payment-status-popup.component';
 
 @Component({
   selector: 'app-program-people-affected',
@@ -26,6 +26,9 @@ export class ProgramPeopleAffectedComponent implements OnInit {
   public thisPhase: ProgramPhase;
   @Output()
   isCompleted: EventEmitter<boolean> = new EventEmitter<boolean>();
+
+  public phaseEnum = ProgramPhase;
+  public userEnum = UserRole;
 
   public program: Program;
   public activePhase: ProgramPhase;
@@ -105,6 +108,16 @@ export class ProgramPeopleAffectedComponent implements OnInit {
       phases: [ProgramPhase.reviewInclusion, ProgramPhase.payment],
       showIfNoValidation: true,
     },
+    {
+      id: BulkActionId.notifyIncluded,
+      enabled: false,
+      label: this.translate.instant(
+        'page.program.program-people-affected.actions.notify-included',
+      ),
+      roles: [UserRole.ProjectOfficer],
+      phases: [ProgramPhase.reviewInclusion, ProgramPhase.payment],
+      showIfNoValidation: true,
+    },
   ];
 
   public submitWarning: any;
@@ -143,6 +156,7 @@ export class ProgramPeopleAffectedComponent implements OnInit {
     const columnDateWidth = 100;
     const columnScoreWidth = 90;
     const columnPhoneNumberWidth = 130;
+
     this.columnsAvailable = [
       {
         prop: 'pa',
@@ -250,6 +264,15 @@ export class ProgramPeopleAffectedComponent implements OnInit {
         width: columnDateTimeWidth,
       },
       {
+        prop: 'notifiedOfInclusion',
+        name: this.translate.instant(
+          'page.program.program-people-affected.column.notified-of-inclusion',
+        ),
+        ...columnDefaults,
+        phases: [ProgramPhase.reviewInclusion, ProgramPhase.payment],
+        width: columnDateTimeWidth,
+      },
+      {
         prop: 'rejected',
         name: this.translate.instant(
           'page.program.program-people-affected.column.rejected',
@@ -265,7 +288,7 @@ export class ProgramPeopleAffectedComponent implements OnInit {
         ),
         ...columnDefaults,
         phases: [ProgramPhase.reviewInclusion, ProgramPhase.payment],
-        width: columnDateTimeWidth,
+        width: 150,
       },
     ];
     this.paymentColumnTemplate = {
@@ -273,6 +296,7 @@ export class ProgramPeopleAffectedComponent implements OnInit {
       name: this.translate.instant(
         'page.program.program-people-affected.column.payment',
       ),
+      installmentIndex: 0,
       ...columnDefaults,
       phases: [ProgramPhase.payment],
       width: columnDateTimeWidth,
@@ -287,10 +311,10 @@ export class ProgramPeopleAffectedComponent implements OnInit {
 
     this.loadColumns();
     if (this.thisPhase === ProgramPhase.payment) {
-      this.addPaymentColumns();
       this.pastTransactions = await this.programsService.getTransactions(
         this.programId,
       );
+      this.addPaymentColumns();
     }
 
     await this.loadData();
@@ -318,13 +342,19 @@ export class ProgramPeopleAffectedComponent implements OnInit {
     );
   }
 
+  private createPaymentColumn(index: number) {
+    const column = JSON.parse(JSON.stringify(this.paymentColumnTemplate)); // Hack to clone without reference;
+    column.name += index;
+    column.prop += index;
+    column.installmentIndex = index;
+    return column;
+  }
+
   private addPaymentColumns() {
     const nrOfInstallments = this.program.distributionDuration;
 
-    for (let p = 0; p < nrOfInstallments; p++) {
-      const column = JSON.parse(JSON.stringify(this.paymentColumnTemplate)); // Hack to clone without reference
-      column.prop += p + 1;
-      column.name += p + 1;
+    for (let index = 1; index <= nrOfInstallments; index++) {
+      const column = this.createPaymentColumn(index);
       this.paymentColumns.push(column);
     }
   }
@@ -426,8 +456,16 @@ export class ProgramPeopleAffectedComponent implements OnInit {
       rejected: person.rejectionDate
         ? formatDate(person.rejectionDate, this.dateFormat, this.locale)
         : null,
+      notifiedOfInclusion: person.inclusionNotificationDate
+        ? formatDate(
+            person.inclusionNotificationDate,
+            this.dateFormat,
+            this.locale,
+          )
+        : null,
       name: person.name,
       phoneNumber: formatPhoneNumber(person.phoneNumber),
+      vnumber: person.vnumber,
       fsp: person.fsp,
     };
 
@@ -436,40 +474,90 @@ export class ProgramPeopleAffectedComponent implements OnInit {
     return personRow;
   }
 
+  private getTransactionOfInstallmentForDid(
+    installmentIndex: number,
+    did: string,
+  ) {
+    return this.pastTransactions.find(
+      (transaction) =>
+        transaction.installment === installmentIndex && transaction.did === did,
+    );
+  }
+
   private fillPaymentColumns(personRow: PersonRow): PersonRow {
-    this.paymentColumns.map((_, index) => {
-      const transactionsThisInstallment = this.pastTransactions.filter(
-        (i) => i.installment === index + 1 && i.did === personRow.did,
+    this.paymentColumns.forEach((paymentColumn) => {
+      const transaction = this.getTransactionOfInstallmentForDid(
+        paymentColumn.installmentIndex,
+        personRow.did,
       );
 
-      if (transactionsThisInstallment.length) {
-        const transaction = transactionsThisInstallment[0];
-        if (transaction.status === 'success') {
-          personRow['payment' + (index + 1)] = formatDate(
-            transaction.installmentdate,
-            this.dateFormat,
-            this.locale,
-          );
-        } else {
-          (personRow['payment' + (index + 1)] = this.translate.instant(
-            'page.program.program-people-affected.failed',
-          )),
-            (personRow['payment' + (index + 1) + '-error'] = transaction.error);
-        }
+      if (!transaction) {
+        return;
       }
+
+      let paymentColumnValue;
+
+      if (transaction.status === 'success') {
+        paymentColumnValue = formatDate(
+          transaction.installmentdate,
+          this.dateFormat,
+          this.locale,
+        );
+      } else if (transaction.status === 'waiting') {
+        paymentColumnValue = this.translate.instant(
+          'page.program.program-people-affected.transaction.waiting',
+        );
+      } else {
+        personRow['payment' + paymentColumn.installmentIndex + '-error'] =
+          transaction.error;
+        paymentColumnValue = this.translate.instant(
+          'page.program.program-people-affected.transaction.failed',
+        );
+      }
+
+      personRow[
+        'payment' + paymentColumn.installmentIndex
+      ] = paymentColumnValue;
     });
     return personRow;
   }
 
-  public async errorPopup(row, column) {
+  public hasVoucherSupport(fsp: string) {
+    const voucherFsps = ['Intersolve-no-whatsapp', 'Intersolve-whatsapp'];
+    return voucherFsps.includes(fsp);
+  }
+
+  public hasError(row: PersonRow, installmentIndex: number) {
+    return !!row['payment' + installmentIndex + '-error'];
+  }
+
+  public async statusPopup(row: PersonRow, column, value: string) {
+    const hasError = this.hasError(row, column.installmentIndex);
+    const content = hasError ? row[column.prop + '-error'] : null;
+    let voucherUrl = null;
+
+    if (this.hasVoucherSupport(row.fsp) && !hasError && !!value) {
+      const voucherBlob = await this.programsService.exportVoucher(
+        row.did,
+        column.installmentIndex,
+      );
+      voucherUrl = window.URL.createObjectURL(voucherBlob);
+    }
+
     const modal: HTMLIonModalElement = await this.modalController.create({
-      component: PaymentErrorPopupComponent,
+      component: PaymentStatusPopupComponent,
       componentProps: {
-        column: column.name,
-        error: row[column.prop + '-error'],
+        title: `${column.name}: ${value}`,
+        content,
+        imageUrl: voucherUrl,
       },
     });
-
+    modal.onDidDismiss().then(() => {
+      // Remove the image from browser memory
+      if (voucherUrl) {
+        window.URL.revokeObjectURL(voucherUrl);
+      }
+    });
     await modal.present();
   }