+++ interfaces/PA-App/_generate-assets-audio.js
@@ -1,3 +1,4 @@
+const dotenv = require('dotenv');
 const readline = require('readline');
 const rl = readline.createInterface({
   input: process.stdin,
@@ -5,6 +6,9 @@ const rl = readline.createInterface({
 });
 const fs = require('fs');
 const { execSync } = require('child_process');
+
+// Load environment-variables from .env file (if available)
+dotenv.config();
 
 // Documentation: https://www.ffmpeg.org/ffmpeg.html
 // Also: https://trac.ffmpeg.org/wiki/Encode/MP3
@@ -27,11 +31,12 @@ function logOutput(error, stdOut, stdErr) {
 }
 
 /**
+ * @param {String} path
  * @param {String} type
  * @returns {Array}
  */
-function getSourceFiles(type) {
-  const files = fs.readdirSync('./').filter((file) => file.match(`${type}$`));
+function getSourceFiles(path, type) {
+  const files = fs.readdirSync(path).filter((file) => file.match(`${type}$`));
 
   if (!files || !files.length) {
     console.warn(`No files found of type: ${type}`);
@@ -44,30 +49,30 @@ function getSourceFiles(type) {
 /**
  * Perform check(s) on source paths/files before proceeding...
  * @param {String} locale
+ * @returns {String} Path to existing source-files
  */
 function checkSourceExists(locale) {
-  const sourcePath = `src/assets/i18n/${locale}/`;
+  const path = `src/assets/i18n/${locale}/`;
 
-  if (!locale || !fs.existsSync(sourcePath)) {
+  if (!locale || !fs.existsSync(path)) {
     console.warn(`No folder available for locale: ${locale}`);
     return process.exit(1);
   }
 
-  // Change to requested folder:
-  process.chdir(sourcePath);
+  return path;
 }
 
 /**
- * Convert audio assets from *.m4a to *.mp3
+ * Convert audio assets from `sourceType` to *.mp3
  * @param {String} locale
+ * @param {String} sourceType
  */
-function convertM4aToMp3(locale) {
-  const sourceFileType = '.m4a';
+function convertToMp3(locale, sourceType) {
+  const sourceFileType = `.${sourceType}`;
   const outputFileType = '.mp3';
 
-  checkSourceExists(locale);
-
-  const sourceFiles = getSourceFiles(sourceFileType);
+  const path = checkSourceExists(locale);
+  const sourceFiles = getSourceFiles(path, sourceFileType);
 
   if (!sourceFiles) {
     return;
@@ -76,10 +81,10 @@ function convertM4aToMp3(locale) {
   sourceFiles.forEach((file) => {
     const outputFile = file.replace(sourceFileType, outputFileType);
 
-    console.log(`Converting to: ${outputFile}`);
+    console.log(`Converting to: ${path}${outputFile}`);
 
     execSync(
-      `${ffmpegPath} -loglevel warning -y -i ${file} -map_metadata -1 -codec:a libmp3lame -q:a 8 ${outputFile}`,
+      `${ffmpegPath} -loglevel warning -y -i ${path}${file} -map_metadata -1 -codec:a libmp3lame -q:a 8 ${path}${outputFile}`,
       logOutput,
     );
   });
@@ -98,8 +103,8 @@ function generateAssetsAudio(locale) {
   const sourceFileType = '.mp3';
   const outputFileType = '.webm';
 
-  checkSourceExists(locale);
-  const sourceFiles = getSourceFiles(sourceFileType);
+  const path = checkSourceExists(locale);
+  const sourceFiles = getSourceFiles(path, sourceFileType);
 
   if (!sourceFiles) {
     return;
@@ -108,10 +113,10 @@ function generateAssetsAudio(locale) {
   sourceFiles.forEach((file) => {
     const outputFile = file.replace(sourceFileType, outputFileType);
 
-    console.log(`Generating: ${outputFile}`);
+    console.log(`Generating: ${path}${outputFile}`);
 
     execSync(
-      `${ffmpegPath} -loglevel warning -y -i ${file} -map_metadata -1  -b:a 64K  -dash 1 ${outputFile}`,
+      `${ffmpegPath} -loglevel warning -y -i ${path}${file} -map_metadata -1  -b:a 64K  -dash 1 ${path}${outputFile}`,
       logOutput,
     );
   });
@@ -122,17 +127,35 @@ function generateAssetsAudio(locale) {
   }, 3000);
 }
 
+function getEnabledLocales(env) {
+  if (!env) return [];
+
+  return env.trim().split(/\s*,\s*/);
+}
+
 ///////////////////////////////////////////////////////////////////////////////
 
-// When the `--convert` flag is used, convert the source-files to mp3 first
-if (process.argv[3] === '--convert') {
-  return convertM4aToMp3(process.argv[2]);
+// When the `--convertFrom` flag is used, convert the source-files to mp3 first
+if (process.argv[3] === '--convertFrom') {
+  return convertToMp3(process.argv[2], process.argv[4]);
 }
 
 // When a locale is provided via the command-line, use that:
 if (process.argv[2]) {
   return generateAssetsAudio(process.argv[2]);
 }
+
+// If locales are defined via ENV-variables, use that:
+if (typeof process.env.NG_LOCALES !== 'undefined') {
+  console.log(`Using locales from ENV: ${process.env.NG_LOCALES}`);
+
+  const audioLocales = getEnabledLocales(process.env.NG_LOCALES);
+
+  return audioLocales.forEach((locale) => {
+    return generateAssetsAudio(locale);
+  });
+}
+
 // Otherwise, ask for it:
 rl.question(
   'For which locale do you want to generate audio assets? : ',