+++ services/121-service/src/cronjob/cronjob.service.ts
@@ -1,10 +1,12 @@
 import { Injectable } from '@nestjs/common';
 import { Cron, CronExpression } from '@nestjs/schedule';
 import { InjectRepository } from '@nestjs/typeorm';
-import { getRepository, LessThan, Repository } from 'typeorm';
+import { getRepository, LessThan, Repository, In, Between, Not } from 'typeorm';
 import { IntersolveBarcodeEntity } from '../programs/fsp/intersolve-barcode.entity';
 import { ProgramEntity } from '../programs/program/program.entity';
 import { WhatsappService } from '../notifications/whatsapp/whatsapp.service';
+import { IntersolveRequestEntity } from '../programs/fsp/intersolve-request.entity';
+import { IntersolveApiService } from '../programs/fsp/api/instersolve.api.service';
 
 @Injectable()
 export class CronjobService {
@@ -12,8 +14,15 @@ export class CronjobService {
   private readonly intersolveBarcodeRepository: Repository<
     IntersolveBarcodeEntity
   >;
+  @InjectRepository(IntersolveRequestEntity)
+  private readonly intersolveRequestRepository: Repository<
+    IntersolveRequestEntity
+  >;
 
-  public constructor(private whatsappService: WhatsappService) {}
+  public constructor(
+    private whatsappService: WhatsappService,
+    private readonly intersolveApiService: IntersolveApiService,
+  ) {}
 
   @Cron(CronExpression.EVERY_DAY_AT_NOON)
   private async cronSendWhatsappReminders(): Promise<void> {
@@ -47,5 +56,49 @@ export class CronjobService {
       `cronSendWhatsappReminders: ${unsentIntersolveBarcodes.length} unsent Intersolve barcodes`,
     );
     console.log('Execution Complete: cronSendWhatsappReminders');
+  }
+
+  @Cron(CronExpression.EVERY_10_MINUTES)
+  private async cronCancelByRefposIntersolve(): Promise<void> {
+    // This function periodically checks if some of the IssueCard calls failed.
+    // and tries to cancel the
+    console.log('Execution Started: cancelByRefposIntersolve');
+
+    const tenMinutes = 10 * 60 * 1000;
+    const tenMinutesAgo = (d => new Date(d.setTime(d.getTime() - tenMinutes)))(
+      new Date(),
+    );
+
+    const twoWeeks = 14 * 24 * 60 * 60 * 1000;
+    const twoWeeksAgo = (d => new Date(d.setTime(d.getTime() - twoWeeks)))(
+      new Date(),
+    );
+    const failedIntersolveRquests = await this.intersolveRequestRepository.find(
+      {
+        where: {
+          updated: Between(twoWeeksAgo, tenMinutesAgo),
+          toCancel: true,
+        },
+      },
+    );
+    for (let intersolveRequest of failedIntersolveRquests) {
+      this.cancelRequestRefpos(intersolveRequest);
+    }
+  }
+
+  private async cancelRequestRefpos(
+    intersolveRequest: IntersolveRequestEntity,
+  ): Promise<void> {
+    intersolveRequest.cancellationAttempts =
+      intersolveRequest.cancellationAttempts + 1;
+    try {
+      const cancelByRefPosResponse = await this.intersolveApiService.cancelTransactionByRefPos(
+        intersolveRequest.refPos,
+      );
+      intersolveRequest.cancelByRefPosResultCode =
+        cancelByRefPosResponse.resultCode;
+    } catch (Error) {
+      console.log('Error cancelling by refpos id', Error, intersolveRequest);
+    }
   }
 }