+++ services/121-service/src/instance/dto/update-instance.dto.ts
@@ -1,0 +1,31 @@
+import { ApiModelProperty } from '@nestjs/swagger';
+import { IsString, IsOptional, IsNotEmpty } from 'class-validator';
+
+export class UpdateInstanceDto {
+  @ApiModelProperty({ example: 'NGO-name' })
+  @IsString()
+  @IsNotEmpty()
+  public readonly name: string;
+
+  @ApiModelProperty({ example: { en: 'NGO display name' } })
+  @IsOptional()
+  public readonly displayName: JSON;
+
+  @ApiModelProperty({
+    example: { en: '<data policy>' },
+  })
+  @IsOptional()
+  public readonly dataPolicy: JSON;
+
+  @ApiModelProperty({
+    example: { en: '<about program>' },
+  })
+  @IsOptional()
+  public readonly aboutProgram: JSON;
+
+  @ApiModelProperty({
+    example: { en: '<contact details>' },
+  })
+  @IsOptional()
+  public readonly contactDetails: JSON;
+}