+++ services/121-service/src/notifications/imagecode/image-code.service.ts
@@ -8,6 +8,7 @@ import crypto from 'crypto';
 import { ImageCodeExportVouchersEntity } from './image-code-export-vouchers.entity';
 import { ConnectionEntity } from '../../sovrin/create-connection/connection.entity';
 import { IntersolveBarcodeEntity } from 'src/programs/fsp/intersolve-barcode.entity';
+import Jimp from 'jimp';
 
 @Injectable()
 export class ImageCodeService {
@@ -22,15 +23,18 @@ export class ImageCodeService {
 
   public constructor() {}
 
-  public async createBarcode(code: string): Promise<string> {
-    const image = await this.generateBarCode(code);
-
+  public async createVoucherUrl(
+    barcodeData: IntersolveBarcodeEntity,
+  ): Promise<string> {
     let barcode = new ImageCodeEntity();
-    barcode.secret = crypto.randomBytes(100).toString('hex') + '.png';
-    barcode.image = image;
+    barcode.secret = crypto.randomBytes(100).toString('hex');
+    barcode.image = await this.generateVoucherImage({
+      dateTime: barcodeData.timestamp,
+      amount: barcodeData.amount,
+      code: barcodeData.barcode,
+      pin: barcodeData.pin,
+    });
     this.imageRepository.save(barcode);
-
-    // Also send a timestamp to see if the secret is older
 
     return EXTERNAL_API.imageCodeUrl + barcode.secret;
   }
@@ -39,32 +43,33 @@ export class ImageCodeService {
     barcodeData: IntersolveBarcodeEntity,
     did: string,
   ): Promise<ImageCodeExportVouchersEntity> {
-    const connection = await this.connectionRepository.findOne({
+    let barcode = new ImageCodeExportVouchersEntity();
+
+    barcode.connection = await this.connectionRepository.findOne({
       where: { did: did },
     });
-    const image = await this.generateBarCode(barcodeData.barcode);
-
-    let barcode = new ImageCodeExportVouchersEntity();
-    barcode.image = image;
-    barcode.connection = connection;
+    barcode.image = await this.generateVoucherImage({
+      dateTime: barcodeData.timestamp,
+      amount: barcodeData.amount,
+      code: barcodeData.barcode,
+      pin: barcodeData.pin,
+    });
     barcode.barcode = barcodeData;
+
     return this.imageExportVouchersRepository.save(barcode);
   }
 
-  public async generateBarCode(code: string): Promise<string> {
-    console.log('code: ', code);
+  private async generateBarCode(code: string): Promise<Buffer> {
     const bwipjs = require('bwip-js');
-    const image = await bwipjs.toBuffer({
-      bcid: 'code128', // Barcode type
-      text: code, // Text to encode
-      scale: 3, // 3x scaling factor
-      height: 10, // Bar height, in millimeters
-      includetext: true, // Show human-readable text
-      textxalign: 'center', // Always good to set this,
+    return await bwipjs.toBuffer({
+      bcid: 'code128',
+      text: code,
+      scale: 2,
+      height: 15,
+      includetext: false,
       backgroundcolor: 'FFFFFF',
       padding: 10,
     });
-    return image;
   }
 
   public async get(secret: string): Promise<any> {
@@ -72,5 +77,65 @@ export class ImageCodeService {
     // Removes the image from the database after getting it
     await this.imageRepository.remove(imageCode);
     return imageCode.image;
+  }
+
+  private formatDate(date: Date): string {
+    const monthNames = [
+      '',
+      'januari',
+      'februari',
+      'maart',
+      'april',
+      'mei',
+      'juni',
+      'juli',
+      'augustus',
+      'september',
+      'oktober',
+      'november',
+      'december',
+    ];
+    const day = String(date.getDate()).padStart(2, '0');
+    const month = monthNames[date.getMonth()];
+
+    return `${day} ${month} ${date.getFullYear()} ${date.getHours()}:${date.getMinutes()}`;
+  }
+
+  public async generateVoucherImage(voucherData: {
+    dateTime: Date;
+    amount: number | string;
+    code: string;
+    pin: string;
+  }): Promise<Buffer> {
+    const voucherBaseFile = './seed-data/voucher/ah-voucher_base.png';
+    const barcodeImage = await this.generateBarCode(voucherData.code);
+
+    // See Jimp documentation: https://www.npmjs.com/package/jimp/v/0.16.1
+    const voucher = await Jimp.read(voucherBaseFile).then(async image => {
+      // Add the generated barcode
+      await Jimp.read(barcodeImage).then(async barcode => {
+        image.composite(barcode, 120, 652);
+      });
+
+      await Jimp.loadFont(Jimp.FONT_SANS_16_BLACK).then(font => {
+        image.print(font, 640, 79, this.formatDate(voucherData.dateTime)); // Date+time in the top-right corner
+        image.print(font, 108, 614, `${voucherData.amount} Albert Heijn`); // Below "Dit ticket is geldig voor", after "€"
+        image.print(font, 640, 604, `Euro ${voucherData.amount}.00`); // Below "Prijs:"
+        image.print(font, 640, 730, voucherData.pin); // Below "Pincode:"
+        image.print(font, 225, 800, voucherData.code); // Barcode numbers
+      });
+
+      // Add a 'scaled' amount to the title of the voucher
+      const titleAmount = new Jimp(50, 50);
+      await Jimp.loadFont(Jimp.FONT_SANS_32_BLACK).then(font => {
+        titleAmount.print(font, 0, 0, voucherData.amount);
+      });
+      titleAmount.scale(0.7);
+      image.composite(titleAmount, 583, 138);
+
+      return image;
+    });
+
+    return await voucher.getBufferAsync(Jimp.MIME_PNG);
   }
 }