+++ services/121-service/src/programs/fsp/api/dto/intersolve-cancel-response.dto.ts
@@ -1,4 +1,4 @@
-import { IntersolveResultCode } from 'src/programs/fsp/api/enum/intersolve-result-code.enum';
+import { IntersolveResultCode } from '../../../../programs/fsp/api/enum/intersolve-result-code.enum';
 
 export class IntersolveCancelResponse {
   public readonly resultCode: IntersolveResultCode;