+++ services/121-service/src/sovrin/credential/credential.service.spec.ts
@@ -36,6 +36,8 @@ import { ImageCodeEntity } from '../../notifications/imagecode/image-code.entity
 import { IntersolveBarcodeEntity } from '../../programs/fsp/intersolve-barcode.entity';
 import { FspAttributeEntity } from '../../programs/fsp/fsp-attribute.entity';
 import { ImageCodeExportVouchersEntity } from '../../notifications/imagecode/image-code-export-vouchers.entity';
+import { IntersolveRequestEntity } from '../../programs/fsp/intersolve-request.entity';
+import { IntersolveInstructionsEntity } from '../../programs/fsp/intersolve-instructions.entity';
 
 describe('CredentialService', (): void => {
   let service: CredentialService;
@@ -140,6 +142,14 @@ describe('CredentialService', (): void => {
             provide: getRepositoryToken(FspAttributeEntity),
             useFactory: repositoryMockFactory,
           },
+          {
+            provide: getRepositoryToken(IntersolveRequestEntity),
+            useFactory: repositoryMockFactory,
+          },
+          {
+            provide: getRepositoryToken(IntersolveInstructionsEntity),
+            useFactory: repositoryMockFactory,
+          },
         ],
       }).compile();
 
@@ -147,7 +157,7 @@ describe('CredentialService', (): void => {
     },
   );
 
-  it('should be defined', (): void => {
+  it('should be defined ', (): void => {
     expect(service).toBeDefined();
   });
 });